reaction_chat_fix

+++ app/src/main/java/com/corona/whatzup/Adapters/MessagesAdapter.java
@@ -111,7 +111,7 @@ public class MessagesAdapter extends RecyclerView.Adapter{
             viewHolder.binding.message.setText(message.getMessage());
 
             if(message.getFeeling() >= 0) {
-                viewHolder.binding.feeling.setImageResource(reactions[(int) message.getFeeling()]);
+                viewHolder.binding.feeling.setImageResource(reactions[message.getFeeling()]);
                 viewHolder.binding.feeling.setVisibility(View.VISIBLE);
             } else {
                 viewHolder.binding.feeling.setVisibility(View.GONE);
@@ -130,7 +130,7 @@ public class MessagesAdapter extends RecyclerView.Adapter{
 
             if(message.getFeeling() >= 0) {
                 //message.setFeeling(reactions[(int) message.getFeeling()]);
-                viewHolder.binding.feeling.setImageResource(reactions[(int) message.getFeeling()]);
+                viewHolder.binding.feeling.setImageResource(reactions[message.getFeeling()]);
                 viewHolder.binding.feeling.setVisibility(View.VISIBLE);
             } else {
                 viewHolder.binding.feeling.setVisibility(View.GONE);
